seeding/requirements.txt
PyQt5>=5.15.0
numpy>=1.23
opencv-python>=4.7.0
Pillow>=9.4.0
PyMuPDF>=1.22.0
ultralytics>=8.0.0
reportlab>=3.6.0
